Update venn.py
Before this patch, the legend did not match the names next to each ellipse. Also, some of the numbers were in the wrong places.

After a bunch of trial and error, I'm pretty this code should display the right numbers in the right places. I suggest you to double check with your own examples..